+++ frb_codegen/src/binary/commands.rs
@@ -1,6 +1,6 @@
 use crate::codegen::ConfigDumpContent;
 use clap::{Args, Parser, Subcommand, ValueEnum};
-use lib_flutter_rust_bridge_codegen::misc::Template;
+use lib_flutter_rust_bridge_codegen::{codegen::RustOpaqueCodecMode, misc::Template};
 use std::path::PathBuf;
 
 // The name `Cli`, `Commands` come from https://docs.rs/clap/latest/clap/_derive/_tutorial/chapter_0/index.html
@@ -145,6 +145,10 @@ pub(crate) struct GenerateCommandArgsPrimary {
     #[arg(long)]
     pub full_dep: bool,
 
+    /// Configure rust opaque implementation, "moi" or "nom"
+    #[arg(long)]
+    pub default_rust_opaque_codec: Option<RustOpaqueCodecModeArg>,
+
     /// Use local version instead of the release version
     #[arg(long, hide = true)]
     pub local: bool,
@@ -226,6 +230,14 @@ pub(crate) enum TemplateArg {
     Plugin,
 }
 
+#[derive(Debug, Copy, Clone, Eq, PartialEq, ValueEnum)]
+pub(crate) enum RustOpaqueCodecModeArg {
+    /// Use moi codec
+    Moi,
+    /// Use nom codec
+    Nom,
+}
+
 #[derive(Debug, Args)]
 pub(crate) struct CreateOrIntegrateCommandCommonArgs {
     /// The name of the generated Rust crate
@@ -265,6 +277,15 @@ impl From<TemplateArg> for Template {
         match value {
             TemplateArg::App => Template::App,
             TemplateArg::Plugin => Template::Plugin,
+        }
+    }
+}
+
+impl From<RustOpaqueCodecModeArg> for RustOpaqueCodecMode {
+    fn from(value: RustOpaqueCodecModeArg) -> Self {
+        match value {
+            RustOpaqueCodecModeArg::Moi => RustOpaqueCodecMode::Moi,
+            RustOpaqueCodecModeArg::Nom => RustOpaqueCodecMode::Nom,
         }
     }
 }